Django_framework/requirements.txt
asgiref==3.8.1
Django==4.2.11
djangorestframework==3.15.1
gunicorn==21.2.0
mysqlclient==2.2.4
packaging==24.0
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
drf_yasg
dj_static
